requirements.txt
APScheduler==3.9.1.post1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cmake==3.25.0
colorama==0.4.6
colorhash==1.2.1
configparser==5.3.0
cycler==0.11.0
Flask==2.2.2
Flask-MonitoringDashboard==3.1.1
fonttools==4.38.0
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.21.6
packaging==23.0
pandas==1.3.5
patsy==0.5.3
Pillow==9.4.0
plotly==5.12.0
plotly-express==0.4.1
psutil==5.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.0.2
scikit-surprise==1.1.1
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
SQLAlchemy==1.4.46
statsmodels==0.13.5
tenacity==8.1.0
threadpoolctl==3.1.0
typing_extensions==4.4.0
tzlocal==2.0.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0